+++ src/backend/notification-service/src/main/java/com/capstone/notificationservice/domain/common/AlarmType.java
@@ -1,5 +1,13 @@
 package com.capstone.notificationservice.domain.common;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
 public enum AlarmType {
-    COMMUNITY, DM
+
+    DM("DM-ALARM"), SERVER("SERVER-ALARM");
+
+    private final String type;
 }
