+++ src/backend/notification-service/src/main/java/com/capstone/notificationservice/domain/dm/dto/DmNotificationDto.java
@@ -2,6 +2,7 @@ package com.capstone.notificationservice.domain.dm.dto;
 
 
 import com.capstone.notificationservice.domain.common.AlarmType;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,6 +14,7 @@ import lombok.NoArgsConstructor;
 public class DmNotificationDto {
     private Long dmRoomId;
     private Long userId;
+    private List<Long> receiverIds;
     private String content;
     private AlarmType alarmType;
 }
