+++ src/backend/notification-service/src/main/java/com/capstone/notificationservice/domain/dm/dto/response/DmNotificationResponse.java
@@ -3,6 +3,7 @@ package com.capstone.notificationservice.domain.dm.dto.response;
 
 import com.capstone.notificationservice.domain.common.AlarmType;
 import com.capstone.notificationservice.domain.dm.entity.Notification;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -14,18 +15,20 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class DmNotificationResponse {
-    Long notificationId;
-    Long dmRoomId;
-    Long userId;
-    String content;
-    AlarmType alarmType;
-    Boolean isRead;
+    private Long notificationId;
+    private Long dmRoomId;
+    private Long userId;
+    private List<Long> receiverIds;
+    private String content;
+    private AlarmType alarmType;
+    private Boolean isRead;
 
-    public static DmNotificationResponse from(Notification notification) {
+    public static DmNotificationResponse from(Notification notification,  List<Long> receiverIds) {
         return DmNotificationResponse.builder()
                 .notificationId(notification.getNotificationId())
                 .dmRoomId(notification.getDmRoomId())
                 .userId(notification.getReceiver().getUserId())
+                .receiverIds(receiverIds)
                 .content(notification.getContent())
                 .alarmType(notification.getAlarmType())
                 .isRead(false)
