+++ src/backend/notification-service/src/main/java/com/capstone/notificationservice/domain/server/dto/MentionType.java
@@ -1,5 +1,13 @@
 package com.capstone.notificationservice.domain.server.dto;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
 public enum MentionType {
-    EVERYONE, HERE, USER
+
+    EVERYONE("MENTION-EVERYONE"), HERE("MENTION-HERE"), USER("MENTION-USER");
+
+    private final String type;
 }
