+++ src/backend/notification-service/src/main/java/com/capstone/notificationservice/domain/user/entity/User.java
@@ -1,5 +1,4 @@
-package com.capstone.notificationservice.global.common.entity;
-
+package com.capstone.notificationservice.domain.user.entity;
 
 import com.capstone.notificationservice.domain.dm.entity.Notification;
 import jakarta.persistence.CascadeType;
@@ -12,7 +11,6 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.PrePersist;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -31,24 +29,12 @@ public class User {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
     private Long userId;
-
     @Column(name = "email", nullable = false, unique = true, length = 50)
     private String email;
-
     @Column(name = "name", nullable = false, length = 50)
     private String username;
-
-    @Column(nullable = false, length = 50)
-    private String nickname;
-
-    @Column(nullable = false, length = 255)
-    private String password;
-
-    @Column(length = 255)
-    private String statusMessage;
     @Column(length = 255)
     private String profileImage;
-    private Date birthdate;
     private Boolean notificationEnabled;
     protected LocalDateTime createdAt;
     protected LocalDateTime modifiedAt;
