+++ src/backend/notification-service/src/main/java/com/capstone/notificationservice/domain/user/repository/userRepository.java
@@ -1,0 +1,7 @@
+package com.capstone.notificationservice.domain.user.repository;
+
+import com.capstone.notificationservice.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface userRepository extends JpaRepository<User, Long> {
+}
